Resize images to target DPI

+++ src/tools.rs
@@ -46,6 +46,6 @@ pub fn load_images_from_archive(archive_path: &Path) -> io::Result<Vec<(String, 
 }
 
 /** Compute a resolution in DPI (PPI actually) from a definition (pixel size) and its printed size (in cm) */
-pub fn compute_dpi(pixel_size: u32, cm_size: f32) -> u32 {
+pub fn compute_dpi(pixel_size: usize, cm_size: f32) -> u32 {
     return (pixel_size as f32 * 2.54 / cm_size) as u32;
 }
